+++ src/tyr/trace_serializer.cc
@@ -84,6 +84,9 @@ json::ArrayPtr serialize_edges(const AttributesController& controller,
       if (edge.has_sac_scale()) {
         edge_map->emplace("sac_scale", static_cast<uint64_t>(edge.sac_scale()));
       }
+      if (edge.has_shoulder()) {
+        edge_map->emplace("shoulder", static_cast<bool>(edge.shoulder()));
+      }
       if (edge.has_sidewalk()) {
         edge_map->emplace("sidewalk", to_string(edge.sidewalk()));
       }