+++ valhalla/proto_conversions.h
@@ -190,6 +190,8 @@ inline TripLeg_Use GetTripLegUse(const baldr::Use use) {
       return TripLeg_Use_kPedestrianUse;
     case baldr::Use::kBridleway:
       return TripLeg_Use_kBridlewayUse;
+    case baldr::Use::kPedestrianCrossing:
+      return TripLeg_Use_kPedestrianCrossingUse;
     case baldr::Use::kRestArea:
       return TripLeg_Use_kRestAreaUse;
     case baldr::Use::kServiceArea: